% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpmnR.R
\docType{package}
\name{bpmnR}
\alias{bpmnR}
\alias{bpmnR-package}
\title{bpmnR}
\description{
Support for BPMN models
}
\author{
\strong{Maintainer}: Gert Janssenswillen \email{gert.janssenswillen@uhasselt.be}

Authors:
\itemize{
  \item Alessio Nigro \email{alessionigro@hotmail.com}
}

Other contributors:
\itemize{
  \item Ivan Esin [contributor]
  \item Hasselt University [copyright holder]
}

}
