% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_dataset_documentation.R
\docType{data}
\name{example_bpmn}
\alias{example_bpmn}
\alias{bpmn_instance}
\title{Example bpmn input for create_bp}
\format{
Eventlog containing 500 patient cases
}
\usage{
bpmn_instance
}
\description{
Contains 5 attributes (tasks, sequenceFlows, gateways, startEvent, endEvent) each one as a dataframe
}
\keyword{datasets}
