% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_bpmn.R
\name{print_bpmn}
\alias{print_bpmn}
\title{Print xml part of bpmn}
\usage{
print_bpmn(x, ...)
}
\arguments{
\item{x}{A bpmn object from create_bpmn function}

\item{...}{Any additional arguments}
}
\value{
No return value, only print model.
}
\description{
Print xml part of bpmn
}
\author{
Alessio Nigro
}
