% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_bpmn.R
\name{render_bpmn-shiny}
\alias{render_bpmn-shiny}
\alias{render_bpmnOutput}
\alias{renderRender_bpmn}
\title{Shiny bindings for render_bpmn}
\usage{
render_bpmnOutput(outputId, width = "100\%", height = "400px")

renderRender_bpmn(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a render_bpmn}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\value{
Rendered BPMN model in Shiny widget.

Rendered BPMN model in Shiny widget.
}
\description{
Output and render functions for using render_bpmn within Shiny
applications and interactive Rmd documents.
}
\author{
Alessio Nigro
}
