% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bpmn.R
\name{write_bpmn}
\alias{write_bpmn}
\alias{write_bpmn.bpmn}
\title{Write XML or HTML to disk.}
\usage{
write_bpmn(bpmn, file, ...)

\method{write_bpmn}{bpmn}(bpmn, file, ...)
}
\arguments{
\item{bpmn}{A BPMN object as a list of data.frames for the BPMN elements and
an XML document for the XML-based interchange format for the BPMN process.}

\item{file}{Path to file or connection to write to.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Writes file to system.
}
\description{
This writes out both XML and normalised HTML. The default behavior will
output the same format which was read. If you want to force output pass
`option = "as_xml"` or `option = "as_html"` respectively.
}
\section{Methods (by class)}{
\itemize{
\item \code{write_bpmn(bpmn)}: Write bpmn to .bpmn file

}}
\author{
Alessio Nigro
}
