% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bpmn.R
\name{create_bpmn}
\alias{create_bpmn}
\title{Create BPMN object.}
\usage{
create_bpmn(tasks, sequenceFlows, gateways, startEvent, endEvent, ...)
}
\arguments{
\item{tasks}{A data.frame of all tasks and their attributes.}

\item{sequenceFlows}{A data.frame of all sequence flows and their attributes.}

\item{gateways}{A data.frame of all gateways and their attributes.}

\item{startEvent}{A data.frame containing the start event and its attributes.}

\item{endEvent}{A data.frame containing the end event and its attributes.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A BPMN object as a list of data.frames for the BPMN elements and an
  XML document for the XML-based interchange format for the BPMN process.
}
\description{
This creates a BPMN object by specifying a set of tasks, sequence flows,
gateways, and a start and end event.
}
\author{
Alessio Nigro
}
