% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xml.R
\name{create_xml}
\alias{create_xml}
\alias{create_xml.bpmn}
\title{Create XML document from BPMN object.}
\usage{
create_xml(bpmn, ...)

\method{create_xml}{bpmn}(bpmn, ...)
}
\arguments{
\item{bpmn}{A BPMN object as a list of data.frames for the BPMN elements.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An XML document for the XML-based interchange format for the BPMN process.
}
\description{
This creates an XML document based on a BPMN object.
}
\author{
Alessio Nigro
}
